pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.8",
    "passlib>=1.7.4",
    "python-jose[cryptography]>=3.4.0",
    "sqlmodel>=0.0.22",
    "uuid>=1.30",
]
